Fix send to eth strange expected types
After debugging the types generation code on the chain side, it is now
clear that for any complicated Msg field (e.g. a Coin) it will have a
new type generated with a derived name and that new type must be
appended to the types. So for a proto field called amount, there will be
a TypeAmount generated for it. If amount is a Coin then TypeAmount will
have the standard denom and amount strings inside of it.

MsgSendToEth was incorrect because it had amount, bridge_fee, and
chain_fee all of type TypeCoin. Instead there are now TypeAmount,
TypeBridgeFee, and TypeChainFee which meet the Coin interface.